feat(sdk): expose runScriptWorkflow() in @agent-relay/sdk/workflows
Lifts the body of `agent-relay run <script>` out of the relay CLI and into
the published @agent-relay/sdk so other tools (ricky in particular) can
import the same .ts/.tsx/.py runner directly instead of spawning the
agent-relay binary.

- New SDK exports from @agent-relay/sdk/workflows:
    runScriptWorkflow, parseTsxStderr, formatWorkflowParseError,
    findLocalSdkWorkspace, ensureLocalSdkWorkflowRuntime, plus
    RunScriptWorkflowOptions / ParsedWorkflowError / LocalSdkWorkspace types.
- @agent-relay/sdk bumped 6.0.2 → 6.1.0 (additive; root dep updated to match).
- relay/src/cli/commands/setup.ts: ~390 lines of helper functions deleted
  and replaced with imports from @agent-relay/sdk/workflows. The CLI's
  `run` command body is unchanged in shape — it just delegates to the SDK
  function. setup.ts also re-exports the helpers so any internal consumer
  importing from setup.ts keeps working.
- 9 unit tests added in packages/sdk/src/workflows/__tests__/run-script.test.ts
  covering the parse-error normalizer, formatter hints, missing-file error,
  unsupported-extension error, and the workspace detector.
