+++ packages/sdk/src/workflows/index.ts
@@ -48,3 +48,16 @@ export * from './proxy-env.js';
 export * from './budget-tracker.js';
 export { applySiblingLinks, buildSiblingLinkScript } from './sibling-links.js';
 export type { SiblingLink, SiblingLinkOptions } from './sibling-links.js';
+export {
+  runScriptWorkflow,
+  parseTsxStderr,
+  formatWorkflowParseError,
+  findLocalSdkWorkspace,
+  ensureLocalSdkWorkflowRuntime,
+} from './run-script.js';
+export type {
+  RunScriptWorkflowOptions,
+  ParsedWorkflowError,
+  LocalSdkWorkspace,
+  ExecFileSyncLike,
+} from './run-script.js';
